+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 6.0.0 (2024-06-20)
+### Breaking Changes
+
+- Type of `SecurityPostureReference.ExcludeExtensions` has been changed from `[]*VirtualMachineExtension` to `[]*string`
+
+### Features Added
+
+- New struct `SecurityPostureReferenceUpdate`
+- New field `ScheduledEventsPolicy` in struct `AvailabilitySetProperties`
+- New field `IsOverridable` in struct `SecurityPostureReference`
+- New field `SecurityPostureReference` in struct `VirtualMachineScaleSetUpdateVMProfile`
+
+
 ## 5.7.0 (2024-04-26)
 ### Features Added